+++ src/main/java/it/eng/idsa/dataapp/web/rest/DataControllerBodyBinary.java
@@ -1,7 +1,8 @@
 package it.eng.idsa.dataapp.web.rest;
 
-import java.util.Optional;
+import java.io.IOException;
 
+import org.apache.http.HttpEntity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -13,17 +14,11 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestPart;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-
-import de.fraunhofer.iais.eis.Message;
 import de.fraunhofer.iais.eis.ContractRequestMessage;
+import de.fraunhofer.iais.eis.Message;
 import de.fraunhofer.iais.eis.RejectionMessage;
 import it.eng.idsa.dataapp.util.MessageUtil;
-import de.fraunhofer.iais.eis.RejectionReason;
-import it.eng.idsa.multipart.builder.MultipartMessageBuilder;
-import it.eng.idsa.multipart.domain.MultipartMessage;
 import it.eng.idsa.multipart.processor.MultipartMessageProcessor;
-import it.eng.idsa.multipart.util.UtilMessageService;
 
 @Controller
 @ConditionalOnProperty(name = "application.dataapp.http.config", havingValue = "mixed")
@@ -41,7 +36,7 @@ public class DataControllerBodyBinary {
 	public ResponseEntity<?> routerBinary(@RequestHeader HttpHeaders httpHeaders,
 			@RequestPart(value = "header") String headerMessage,
 			@RequestHeader(value = "Response-Type", required = false) String responseType,
-			@RequestPart(value = "payload", required = false) String payload) throws JsonProcessingException {
+			@RequestPart(value = "payload", required = false) String payload) throws IOException {
 
 		logger.info("Multipart/mixed request");
 
@@ -65,25 +60,23 @@ public class DataControllerBodyBinary {
 		} 
 		if(responsePayload == null && message instanceof ContractRequestMessage) {
 			logger.info("Creating rejection message since contract agreement was not found");
-			headerResponse = UtilMessageService.getRejectionMessage(RejectionReason.NOT_FOUND);
+			headerResponse = messageUtil.createRejectionCommunicationLocalIssues(message);
+//					.getRejectionMessage(RejectionReason.NOT_FOUND);
 		}	
 		
 		if (responsePayload != null && responsePayload.contains("ids:rejectionReason")) {
 			headerResponse = MultipartMessageProcessor.getMessage(responsePayload);
 			responsePayload = null;
 		}
-		MultipartMessage responseMessage = new MultipartMessageBuilder()
-				.withHeaderContent(headerResponse)
-				.withPayloadContent(responsePayload)
-				.build();
-		String responseMessageString = MultipartMessageProcessor.multipartMessagetoString(responseMessage, false, Boolean.TRUE);
-		
-		Optional<String> boundary = MultipartMessageProcessor.getMessageBoundaryFromMessage(responseMessageString);
-		String contentType = "multipart/mixed; boundary=" + boundary.orElse("---aaa") + ";charset=UTF-8";
 
+		HttpEntity resultEntity = messageUtil.createMultipartMessageForm(
+				MultipartMessageProcessor.serializeToJsonLD(headerResponse),
+				responsePayload);
+		String contentType = resultEntity.getContentType().getValue();
+		contentType = contentType.replace("multipart/form-data", "multipart/mixed");
 		return ResponseEntity.ok()
 				.header("foo", "bar")
 				.contentType(MediaType.parseMediaType(contentType))
-				.body(responseMessageString);
+				.body(resultEntity.getContent().readAllBytes());
 	}
 }