+++ src/main/java/it/eng/idsa/dataapp/web/rest/DataControllerBodyForm.java
@@ -5,7 +5,6 @@ import java.io.IOException;
 import java.util.Base64;
 
 import org.apache.http.HttpEntity;
-import org.apache.http.entity.ContentType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -21,10 +20,8 @@ import org.springframework.web.multipart.MultipartFile;
 import de.fraunhofer.iais.eis.ContractRequestMessage;
 import de.fraunhofer.iais.eis.Message;
 import de.fraunhofer.iais.eis.RejectionMessage;
-import de.fraunhofer.iais.eis.RejectionReason;
 import it.eng.idsa.dataapp.util.MessageUtil;
 import it.eng.idsa.multipart.processor.MultipartMessageProcessor;
-import it.eng.idsa.multipart.util.UtilMessageService;
 
 @RestController
 @ConditionalOnProperty(name = "application.dataapp.http.config", havingValue = "form")
@@ -78,11 +75,12 @@ public class DataControllerBodyForm {
 		Message headerResponse = messageUtil.getResponseHeader(message);
 		String responsePayload = null;
 		if (!(headerResponse instanceof RejectionMessage)) {
-			responsePayload = messageUtil.createResponsePayload(message, payload.toString());
+			responsePayload = messageUtil.createResponsePayload(message, payload != null ? payload.toString() : null);
 		}
 		if(responsePayload == null && message instanceof ContractRequestMessage) {
 			logger.info("Creating rejection message since contract agreement was not found");
-			headerResponse = UtilMessageService.getRejectionMessage(RejectionReason.NOT_FOUND);
+//			headerResponse = UtilMessageService.getRejectionMessage(RejectionReason.NOT_FOUND);
+			headerResponse = messageUtil.createRejectionCommunicationLocalIssues(message);
 		}		
 		if (responsePayload != null && responsePayload.contains("ids:rejectionReason")) {
 			headerResponse = MultipartMessageProcessor.getMessage(responsePayload);
@@ -92,9 +90,7 @@ public class DataControllerBodyForm {
 		// prepare body response - multipart message.
 		HttpEntity resultEntity = messageUtil.createMultipartMessageForm(
 				MultipartMessageProcessor.serializeToJsonLD(headerResponse),
-				responsePayload,
-				null,
-				ContentType.APPLICATION_JSON);
+				responsePayload);
 		
 		return ResponseEntity.ok()
 				.header("foo", "bar")